split up into more files to make it possible to use some animations on M0
