+++ examples/led_animation_simpletest.py
@@ -4,11 +4,13 @@ interval.
 
 For NeoPixel FeatherWing. Update pixel_pin and pixel_num to match your wiring if using
 a different form of NeoPixels.
+
+This example does not work on SAMD21 (M0) boards.
 """
 import board
 import neopixel
 from adafruit_led_animation.animation import Comet, Chase
-from adafruit_led_animation.helper import AnimationSequence
+from adafruit_led_animation.sequence import AnimationSequence
 from adafruit_led_animation.color import PURPLE, WHITE
 
 # Update to match the pin connected to your NeoPixels
